moving the fixture to common location and adding fspath check
